# openapi.yaml
openapi: 3.2.0
info:
  title: GREENHERB API
  description: Plataforma Web de Gestão Inteligente de uma Estufa
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /planos:
    get:
      tags:
        - Planos de Cultivo
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
    post:
      tags:
        - Planos de Cultivo
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanoCultivo"
      responses:
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /planos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      tags:
        - Planos de Cultivo
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanoCultivo"
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
...
  /tarefas:
    get:
      tags:
        - Tarefas Operacionais
      security:
        - cookieAuth: []
      parameters:
        - name: loteId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      tags:
        - Tarefas Operacionais
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tarefa"
      responses:
        "201":
          description: Created
  /tarefas/exportar:
    get:
      tags:
        - Tarefas Operacionais
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /tarefas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      tags:
        - Tarefas Operacionais
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Tarefa"
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access
  schemas:
...
    PlanoCultivo:
      type: object
      required: [tipo]
      properties:
        tipo:
          type: string
          enum: [Regular, Emergência, Pontual]
        ervaId:
          type: string
        condicoesIdeais:
          type: object
          properties:
            temperaturaMin:
              type: number
            temperaturaMax:
              type: number
            humidadeMin:
              type: number
            humidadeMax:
              type: number
            luminosidadeMin:
              type: number
            luminosidadeMax:
              type: number
        intervencoes:
          type: array
          items:
            type: object
            properties:
              tipo:
                type: string
              frequenciaHoras:
                type: number
              dosagem:
                type: string
        autorizadoPor:
          type: string
...
    Tarefa:
      type: object
      required: [loteId, tipo, dataAgendada]
      properties:
        loteId:
          type: string
        tipo:
          type: string
          enum: [Rega, Fertilização, Colheita, Monitorização, Outro]
        estado:
          type: string
          enum: [Pendente, Executada, Expirada]
        dataAgendada:
          type: string
          format: date-time
        dataExecucao:
          type: string
          format: date-time
        responsavelId:
          type: string
...